refactor series reader to make it small
